refactor: move shared code to plugin-kubernetes-lib
Connection, OAuthTokenProvider, SideCar, ClientService, PodService,
PodLogService, InstanceService, LoggingOutputStream, watchers, and the
containerToMap/deepMerge/moveFile/mapSidecarResources helpers now come
from io.kestra.plugin:plugin-kubernetes-lib, shared with
plugin-ee-kubernetes. Also bumps fabric8 kubernetes-client 7.7.0 to
7.8.0 via the lib.
