+++ build.gradle
@@ -56,8 +56,8 @@ dependencies {
     compileOnly group: "io.kestra", name: "core", version: kestraVersion
 
     // plugins
-    api 'io.fabric8:kubernetes-client:7.8.0'
-    api group: 'org.apache.commons', name: 'commons-compress'
+    // Shared kernel with plugin-ee-kubernetes; brings io.fabric8:kubernetes-client and commons-compress transitively.
+    api 'io.kestra.plugin:plugin-kubernetes-lib:0.1.0-SNAPSHOT'
 }
