+++ src/main/java/io/kestra/plugin/kubernetes/kubectl/Get.java
@@ -25,7 +25,7 @@ import io.kestra.core.serializers.FileSerde;
 import io.kestra.plugin.kubernetes.AbstractPod;
 import io.kestra.plugin.kubernetes.models.Metadata;
 import io.kestra.plugin.kubernetes.models.ResourceStatus;
-import io.kestra.plugin.kubernetes.services.PodService;
+import io.kestra.plugin.kubernetes.shared.services.PodService;
 import io.kestra.plugin.kubernetes.services.ResourceWaitService;
 
 import io.fabric8.kubernetes.client.KubernetesClientException;
