+++ src/main/java/io/kestra/plugin/kubernetes/kubectl/Patch.java
@@ -12,7 +12,7 @@ import io.kestra.plugin.kubernetes.AbstractPod;
 import io.kestra.plugin.kubernetes.models.Metadata;
 import io.kestra.plugin.kubernetes.models.PatchStrategy;
 import io.kestra.plugin.kubernetes.models.ResourceStatus;
-import io.kestra.plugin.kubernetes.services.PodService;
+import io.kestra.plugin.kubernetes.shared.services.PodService;
 import io.kestra.plugin.kubernetes.services.ResourceWaitService;
 
 import io.fabric8.kubernetes.client.dsl.base.PatchContext;
