--- src/main/java/io/kestra/plugin/kubernetes/services/PodService.java
@@ -1,573 +1,0 @@
-package io.kestra.plugin.kubernetes.services;
-
-import java.io.ByteArrayOutputStream;
-import java.io.File;
-import java.io.IOException;
-import java.nio.charset.StandardCharsets;
-import java.nio.file.Path;
-import java.time.Duration;
-import java.util.ArrayList;
-import java.util.Comparator;
-import java.util.List;
-import java.util.Optional;
-import java.util.concurrent.TimeUnit;
-import java.util.concurrent.TimeoutException;
-import java.util.concurrent.atomic.AtomicInteger;
-import java.util.function.Predicate;
-
-import org.slf4j.Logger;
-
-import io.kestra.core.exceptions.IllegalVariableEvaluationException;
-import io.kestra.core.models.tasks.retrys.Exponential;
-import io.kestra.core.models.tasks.runners.AbstractLogConsumer;
-import io.kestra.core.runners.RunContext;
-import io.kestra.core.utils.RetryUtils;
-import io.kestra.plugin.kubernetes.models.Connection;
-
-import io.fabric8.kubernetes.api.model.*;
-import io.fabric8.kubernetes.client.Config;
-import io.fabric8.kubernetes.client.KubernetesClient;
-import io.fabric8.kubernetes.client.KubernetesClientException;
-import io.fabric8.kubernetes.client.KubernetesClientTimeoutException;
-import io.fabric8.kubernetes.client.dsl.ContainerResource;
-import io.fabric8.kubernetes.client.dsl.PodResource;
-
-abstract public class PodService {
-    private static final List<String> COMPLETED_PHASES = List.of("Succeeded", "Failed", "Unknown"); // see https://kubernetes.io/docs/concepts/workloads/pods/pod-lifecycle/#pod-phase
-    private static final String SIDECAR_FILES_CONTAINER_NAME = "out-files";
-    // 0 means "don't wait at all"; 30s gives the pod time to become ready before exec connections are established
-    public static final int EXEC_READY_WAIT_TIMEOUT_MS = 30_000;
-
-    private static final int UPLOAD_RETRY_MAX_ATTEMPTS = 5;
-
-    public static KubernetesClient client(RunContext runContext, Connection connection) throws IllegalVariableEvaluationException {
-        return client(runContext, connection, false);
-    }
-
-    public static KubernetesClient client(RunContext runContext, Connection connection, boolean inheritClusterConfig) throws IllegalVariableEvaluationException {
-        return connection != null ? client(connection.toConfig(runContext, inheritClusterConfig)) : client(null);
-    }
-
-    public static KubernetesClient client(Config config) {
-        return config != null ? ClientService.of(config) : ClientService.of();
-    }
-
-    public static Pod waitForInitContainerRunning(KubernetesClient client, Pod pod, String container, Duration waitUntilRunning) {
-        return PodService.podRef(client, pod)
-            .waitUntilCondition(
-                j -> j != null &&
-                    j.getStatus() != null &&
-                    j.getStatus().getInitContainerStatuses() != null &&
-                    j.getStatus()
-                        .getInitContainerStatuses()
-                        .stream()
-                        .filter(containerStatus -> containerStatus.getName().equals(container))
-                        .anyMatch(containerStatus -> containerStatus.getState() != null && containerStatus.getState().getRunning() != null),
-                waitUntilRunning.toSeconds(),
-                TimeUnit.SECONDS
-            );
-    }
-
-    public static Pod waitForPodReady(KubernetesClient client, Pod pod, Duration waitUntilRunning) {
-        boolean hasSidecar = pod.getSpec().getContainers().stream()
-            .anyMatch(c -> SIDECAR_FILES_CONTAINER_NAME.equals(c.getName()));
-
-        return PodService.podRef(client, pod)
-            .waitUntilCondition(
-                j -> j != null &&
-                    j.getStatus() != null && (PodPhase.FAILED.value().equals(j.getStatus().getPhase()) ||
-                        (j.getStatus().getContainerStatuses() != null &&
-                            j.getStatus().getContainerStatuses().stream()
-                                .anyMatch(
-                                    cs -> cs.getState() != null &&
-                                        cs.getState().getWaiting() != null &&
-                                        cs.getState().getWaiting().getReason() != null &&
-                                        !TransientWaitingReason.contains(cs.getState().getWaiting().getReason())
-                                ))
-                        ||
-                        j.getStatus()
-                            .getConditions()
-                            .stream()
-                            .anyMatch(
-                                podCondition -> ("ContainersReady".equals(podCondition.getType()) &&
-                                    (hasSidecar || "True".equals(podCondition.getStatus()))) ||
-                                    ("PodCompleted".equals(podCondition.getReason()))
-                            )),
-                waitUntilRunning.toSeconds(),
-                TimeUnit.SECONDS
-            );
-    }
-
-    public static boolean isCompletedPhase(String phase) {
-        return phase != null && COMPLETED_PHASES.contains(phase);
-    }
-
-    private static boolean isStartedOrCompleted(Pod pod) {
-        if (pod == null || pod.getStatus() == null) {
-            return false;
-        }
-        var phase = pod.getStatus().getPhase();
-        if (isCompletedPhase(phase)) {
-            return true;
-        }
-        return PodPhase.RUNNING.value().equals(phase) &&
-            pod.getStatus().getContainerStatuses() != null &&
-            pod.getStatus().getContainerStatuses().stream()
-                .anyMatch(c -> c.getState() != null && c.getState().getRunning() != null);
-    }
-
-    public static Pod waitForContainersStartedOrCompleted(KubernetesClient client, Logger logger, Pod pod, Duration waitUntilRunning) {
-        var podResource = podRef(client, pod);
-
-        // Fast path: pod already satisfies the condition (e.g. on resume with a Running/Succeeded pod)
-        var current = podResource.get();
-        if (current == null) {
-            throw new KubernetesClientException("Pod was deleted while waiting for containers to start: " + pod.getMetadata().getName());
-        }
-        if (isStartedOrCompleted(current)) {
-            return current;
-        }
-
-        var remaining = waitUntilRunning;
-        var chunk = Duration.ofMinutes(5);
-
-        while (remaining.toSeconds() > 0) {
-            var waitTime = remaining.compareTo(chunk) < 0 ? remaining : chunk;
-            remaining = remaining.minus(waitTime);
-
-            try {
-                var result = podResource.waitUntilCondition(
-                    PodService::isStartedOrCompleted,
-                    waitTime.toSeconds(),
-                    TimeUnit.SECONDS
-                );
-                if (result != null) {
-                    return result;
-                }
-            } catch (KubernetesClientTimeoutException e) {
-                // chunk expired — fall through to GET check
-            } catch (KubernetesClientException e) {
-                logger.debug("Watch error while waiting for containers to start after {}s, rechecking pod state", waitTime.toSeconds(), e);
-            }
-
-            podResource = podRef(client, pod);
-            var polled = podResource.get();
-            if (polled == null) {
-                throw new KubernetesClientException("Pod was deleted while waiting for containers to start: " + pod.getMetadata().getName());
-            }
-            if (isStartedOrCompleted(polled)) {
-                return polled;
-            }
-        }
-
-        throw new KubernetesClientTimeoutException(pod, waitUntilRunning.toSeconds(), TimeUnit.SECONDS);
-    }
-
-    public static Pod waitForCompletionExcept(KubernetesClient client, Logger logger, Pod pod, Duration waitRunning, String except) {
-        return waitForCompletion(
-            client,
-            logger,
-            pod,
-            waitRunning,
-            j -> j != null &&
-                j.getStatus() != null &&
-                j.getStatus().getContainerStatuses() != null &&
-                j.getStatus()
-                    .getContainerStatuses()
-                    .stream()
-                    .anyMatch(containerStatus -> !containerStatus.getName().equals(except))
-                &&
-                j.getStatus()
-                    .getContainerStatuses()
-                    .stream()
-                    .filter(containerStatus -> !containerStatus.getName().equals(except))
-                    .allMatch(containerStatus -> containerStatus.getState() != null && containerStatus.getState().getTerminated() != null)
-        );
-    }
-
-    public static Pod waitForCompletion(KubernetesClient client, Logger logger, Pod pod, Duration waitRunning) {
-        return waitForCompletion(
-            client,
-            logger,
-            pod,
-            waitRunning,
-            j -> j != null &&
-                j.getStatus() != null &&
-                COMPLETED_PHASES.contains(j.getStatus().getPhase())
-        );
-    }
-
-    public static Pod waitForCompletion(KubernetesClient client, Logger logger, Pod pod, Duration waitRunning, Predicate<Pod> condition) {
-        var podResource = podRef(client, pod);
-        var remaining = waitRunning;
-        var chunk = Duration.ofMinutes(5);
-
-        while (remaining.toSeconds() > 0) {
-            var waitTime = remaining.compareTo(chunk) < 0 ? remaining : chunk;
-            remaining = remaining.minus(waitTime);
-
-            try {
-                var ended = podResource.waitUntilCondition(condition, waitTime.toSeconds(), TimeUnit.SECONDS);
-                if (ended != null) {
-                    return ended;
-                }
-            } catch (KubernetesClientTimeoutException e) {
-                // chunk expired — fall through to GET check below
-            } catch (KubernetesClientException e) {
-                // watch error — fall through to GET check below
-                logger.debug("Watch error while waiting for pod completion, checking pod state", e);
-            }
-
-            podResource = podRef(client, pod);
-            var current = podResource.get();
-            if (current == null) {
-                var podName = pod.getMetadata().getName();
-                logger.warn("Pod '{}' was deleted before reaching a terminal phase", podName);
-                throw new KubernetesClientException("Pod was deleted before reaching a terminal phase: " + podName);
-            }
-            if (condition.test(current)) {
-                return current;
-            }
-        }
-
-        throw new KubernetesClientTimeoutException(pod, waitRunning.toSeconds(), TimeUnit.SECONDS);
-    }
-
-    public static IllegalStateException failedMessage(Pod pod) throws IllegalStateException {
-        if (pod.getStatus() == null) {
-            return new IllegalStateException("Pods terminated without any status !");
-        }
-
-        return (pod.getStatus().getContainerStatuses() == null ? new ArrayList<ContainerStatus>() : pod.getStatus().getContainerStatuses())
-            .stream()
-            .filter(containerStatus -> containerStatus.getState() != null && containerStatus.getState().getTerminated() != null)
-            .map(containerStatus -> containerStatus.getState().getTerminated())
-            .findFirst()
-            .map(
-                containerStateTerminated -> new IllegalStateException(
-                    "Pods terminated with status '" + pod.getStatus().getPhase() + "', " +
-                        "exitcode '" + containerStateTerminated.getExitCode() + "' & " +
-                        "message '" + containerStateTerminated.getMessage() + "'"
-                )
-            )
-            .orElseGet(() ->
-            {
-                if (pod.getStatus().getContainerStatuses() != null) {
-                    Optional<String> waitingReason = pod.getStatus().getContainerStatuses().stream()
-                        .filter(cs -> cs.getState() != null && cs.getState().getWaiting() != null)
-                        .map(cs -> cs.getState().getWaiting().getReason())
-                        .filter(reason -> reason != null && !TransientWaitingReason.contains(reason))
-                        .findFirst();
-                    if (waitingReason.isPresent()) {
-                        return new IllegalStateException("Pod failed before container start: " + waitingReason.get());
-                    }
-                }
-                return new IllegalStateException("Pod failed with phase '" + pod.getStatus().getPhase() + "'");
-            });
-    }
-
-    public static void checkContainerFailures(Pod pod, String exceptContainer, Logger logger) throws IllegalStateException {
-        if (pod.getStatus() == null || pod.getStatus().getContainerStatuses() == null) {
-            return;
-        }
-
-        pod.getStatus().getContainerStatuses().stream()
-            .filter(containerStatus -> !containerStatus.getName().equals(exceptContainer))
-            .filter(containerStatus -> containerStatus.getState() != null && containerStatus.getState().getTerminated() != null)
-            .filter(containerStatus -> containerStatus.getState().getTerminated().getExitCode() != 0)
-            .findFirst()
-            .ifPresent(containerStatus ->
-            {
-                ContainerStateTerminated terminated = containerStatus.getState().getTerminated();
-                String errorMsg = "Container '" + containerStatus.getName() + "' failed with exit code " +
-                    terminated.getExitCode() +
-                    (terminated.getReason() != null ? ", reason: " + terminated.getReason() : "") +
-                    (terminated.getMessage() != null ? ", message: " + terminated.getMessage() : "");
-
-                logger.error(errorMsg);
-                throw new IllegalStateException(errorMsg);
-            });
-    }
-
-    public static PodResource podRef(KubernetesClient client, Pod pod) {
-        return client.pods()
-            .inNamespace(pod.getMetadata().getNamespace())
-            .withName(pod.getMetadata().getName());
-    }
-
-    /**
-     * Retry file operations with exponential backoff optimized for freshly provisioned nodes.
-     */
-    public static Boolean withRetries(Logger logger, String where, RetryUtils.CheckedSupplier<Boolean> call) throws IOException {
-        var attempt = new AtomicInteger(0);
-        try {
-            return RetryUtils.Instance.<Boolean, IOException> builder()
-                .policy(
-                    Exponential.builder()
-                        .delayFactor(2.0)
-                        .interval(Duration.ofSeconds(1))
-                        .maxInterval(Duration.ofSeconds(10))
-                        .maxDuration(Duration.ofSeconds(60))
-                        .maxAttempts(UPLOAD_RETRY_MAX_ATTEMPTS)
-                        .build()
-                )
-                .logger(logger)
-                // On exhaustion, Failsafe's Fallback throws RetryUtils.RetryFailed rather than an IOException,
-                // which would otherwise discard the detailed hint message thrown below. Unwrap it back to the
-                // last real cause so that detail survives all the way to the caller instead of being replaced
-                // by a bare "Failed to call" message.
-                .failureFunction(retryFailed -> retryFailed.getCause() instanceof IOException ioException
-                    ? ioException
-                    : new IOException("Failed to call '" + where + "'", retryFailed.getCause()))
-                .build()
-                // Stay on the result-predicate overload (rather than run(IOException.class, ...), which would
-                // narrow Failsafe's default "retry on any exception" to IOException only): fabric8 calls can
-                // throw KubernetesClientException and other non-IOException runtime errors that must still be
-                // retried the same way a bare `false` result is.
-                .run(
-                    object -> !object,
-                    () ->
-                    {
-                        var currentAttempt = attempt.incrementAndGet();
-
-                        if (!call.get()) {
-                            // fabric8's exec/copy calls return a bare boolean and swallow the underlying exec/tar
-                            // error, so there is no exception or stderr to surface here. Throwing here — instead
-                            // of just returning false, as before — makes this failure the retry policy's tracked
-                            // "last exception", so the detailed hint below is what the failureFunction above
-                            // unwraps on exhaustion, rather than a bare RetryFailed with no cause to unwrap.
-                            throw new IOException(
-                                "Failed to call '" + where + "' after " + currentAttempt + "/" + UPLOAD_RETRY_MAX_ATTEMPTS +
-                                    " attempts — the Kubernetes copy/exec call kept reporting failure without further detail; " +
-                                    "verify connectivity to the pod and that the file-sidecar image provides a working 'sh' and 'tar'"
-                            );
-                        }
-
-                        return true;
-                    }
-                );
-        } catch (IOException e) {
-            throw e;
-        } catch (Throwable e) {
-            throw new IOException("Failed to call '" + where + "'", e);
-        }
-    }
-
-    /**
-     * Runs a command inside a container and returns its stdout, but only if it exits with code 0.
-     * Used to cross-check fabric8 copy/upload results, which can report success even when the
-     * underlying tar transfer was truncated.
-     * <p>
-     * On timeout, only the client-side {@code ExecWatch} is closed (via try-with-resources); fabric8's exec
-     * API has no way to signal the remote process, so it may keep running inside the (short-lived,
-     * verification-only) sidecar container until that container is torn down with the rest of the pod.
-     * <p>
-     * Every fabric8 {@code exec} call re-waits for the whole pod to report Ready (up to
-     * {@link #EXEC_READY_WAIT_TIMEOUT_MS}) before opening the exec connection — see
-     * {@code PodOperationsImpl#getURL}. That wait is structurally doomed to run to its full timeout here:
-     * the pod can't be Ready while the init/sidecar file-transfer containers are still blocked on marker
-     * files, which is exactly when upload verification runs. Since this call always follows an already-successful
-     * exec on the very same container (the upload itself), the container is proven reachable, so skip that
-     * redundant wait entirely instead of paying it again on every verification call.
-     */
-    public static Optional<String> execOutput(ContainerResource container, Logger logger, Duration timeout, String... command) {
-        var output = new ByteArrayOutputStream();
-        try (var watch = container.withReadyWaitTimeout(0).writingOutput(output).exec(command)) {
-            var exitCode = watch.exitCode().get(timeout.toSeconds(), TimeUnit.SECONDS);
-            if (exitCode == null || exitCode != 0) {
-                return Optional.empty();
-            }
-            return Optional.of(output.toString(StandardCharsets.UTF_8));
-        } catch (InterruptedException e) {
-            Thread.currentThread().interrupt();
-            return Optional.empty();
-        } catch (TimeoutException e) {
-            // Best-effort: a timeout here is otherwise indistinguishable from "verification passed", so
-            // log it explicitly rather than silently skipping the check like the generic catch below.
-            logger.info(
-                "Verification command '{}' timed out after {}s in the file-sidecar container, skipping upload verification",
-                String.join(" ", command), timeout.toSeconds()
-            );
-            return Optional.empty();
-        } catch (Exception e) {
-            logger.debug("Verification command '{}' could not be executed in the file-sidecar container", String.join(" ", command), e);
-            return Optional.empty();
-        }
-    }
-
-    /**
-     * Retries a fallible verification check (e.g. a per-file upload size/count comparison), applying the
-     * same retry treatment given to the raw upload/exec calls above so a single transient exec failure
-     * during verification doesn't turn an otherwise-good upload into a hard task failure.
-     */
-    public static void withVerificationRetries(Logger logger, String where, CheckedRunnable check) throws IOException {
-        try {
-            RetryUtils.Instance.<Void, IOException> builder()
-                .policy(
-                    Exponential.builder()
-                        .delayFactor(2.0)
-                        .interval(Duration.ofSeconds(1))
-                        .maxInterval(Duration.ofSeconds(10))
-                        .maxDuration(Duration.ofSeconds(60))
-                        .maxAttempts(UPLOAD_RETRY_MAX_ATTEMPTS)
-                        .build()
-                )
-                .logger(logger)
-                // On exhaustion, Failsafe's Fallback throws RetryUtils.RetryFailed rather than the checked
-                // exception the policy handles, which would otherwise discard the actionable message built by
-                // verifyUpload (e.g. "expected N file(s) but found M"). Unwrap it back to the last real cause
-                // so that detail survives all the way to the caller instead of being replaced by a bare
-                // "Failed to call" message.
-                .failureFunction(retryFailed -> retryFailed.getCause() instanceof IOException ioException
-                    ? ioException
-                    : new IOException("Failed to call '" + where + "'", retryFailed.getCause()))
-                .build()
-                .run(
-                    IOException.class,
-                    () ->
-                    {
-                        check.run();
-                        return null;
-                    }
-                );
-        } catch (IOException e) {
-            throw e;
-        } catch (Throwable e) {
-            throw new IOException("Failed to call '" + where + "'", e);
-        }
-    }
-
-    @FunctionalInterface
-    public interface CheckedRunnable {
-        void run() throws IOException;
-    }
-
-    public static void uploadMarker(RunContext runContext, PodResource podResource, Logger logger, String marker, String container) throws IOException {
-        File markerFile = tempDir(runContext).resolve(marker).toFile();
-        if (!markerFile.createNewFile()) {
-            throw new IOException("Unable to create the marker file: " + markerFile.getAbsolutePath());
-        }
-
-        withRetries(
-            logger,
-            "uploadMarker",
-            () -> podResource
-                .inContainer(container)
-                .withReadyWaitTimeout(EXEC_READY_WAIT_TIMEOUT_MS)
-                .file("/kestra/" + marker)
-                .upload(markerFile.toPath())
-        );
-
-        if (!markerFile.delete()) {
-            logger.debug("Unable to delete the marker file: {}", markerFile.getAbsolutePath());
-        }
-        logger.debug(marker + " marker uploaded");
-    }
-
-    /**
-     * Fetch and log Kubernetes pod events (e.g. FailedScheduling, Evicted, ImagePullBackOff).
-     */
-    public static void logPodEvents(KubernetesClient client, Pod pod, Logger logger, AbstractLogConsumer logConsumer) {
-        if (pod == null || pod.getMetadata() == null) {
-            return;
-        }
-
-        String namespace = pod.getMetadata().getNamespace();
-        String podName = pod.getMetadata().getName();
-
-        try {
-            client.v1().events()
-                .inNamespace(namespace)
-                .withField("involvedObject.name", podName)
-                .list()
-                .getItems()
-                .stream()
-                .filter(event -> "Warning".equals(event.getType()))
-                .sorted(
-                    Comparator.comparing(
-                        Event::getLastTimestamp,
-                        Comparator.nullsLast(Comparator.naturalOrder())
-                    )
-                )
-                .forEach(event ->
-                {
-                    String reason = event.getReason() == null ? "" : event.getReason();
-                    String message = event.getMessage() == null ? "" : event.getMessage();
-
-                    logConsumer.accept(
-                        "[pod-event] " + reason + " - " + message,
-                        true
-                    );
-                });
-
-        } catch (Exception e) {
-            logger.warn("Failed to fetch events for pod '{}'", podName, e);
-        }
-    }
-
-    public static boolean hasAnyContainerStarted(Pod pod) {
-        if (pod.getStatus() == null || pod.getStatus().getContainerStatuses() == null) {
-            return false;
-        }
-        return pod.getStatus().getContainerStatuses().stream()
-            .anyMatch(
-                cs -> cs.getState() != null &&
-                    (cs.getState().getRunning() != null || cs.getState().getTerminated() != null)
-            );
-    }
-
-    public static boolean hasNonTransientWaitingContainer(Pod pod) {
-        if (pod.getStatus() == null || pod.getStatus().getContainerStatuses() == null) {
-            return false;
-        }
-        return pod.getStatus().getContainerStatuses().stream()
-            .anyMatch(
-                cs -> cs.getState() != null &&
-                    cs.getState().getWaiting() != null &&
-                    cs.getState().getWaiting().getReason() != null &&
-                    !TransientWaitingReason.contains(cs.getState().getWaiting().getReason())
-            );
-    }
-
-    public static Path tempDir(RunContext runContext) {
-        return runContext.workingDir().path().resolve("working-dir");
-    }
-
-    public enum TransientWaitingReason {
-        CONTAINER_CREATING("ContainerCreating"),
-        POD_INITIALIZING("PodInitializing");
-
-        private final String reason;
-
-        TransientWaitingReason(String reason) {
-            this.reason = reason;
-        }
-
-        public static boolean contains(String reason) {
-            for (TransientWaitingReason r : values()) {
-                if (r.reason.equals(reason)) {
-                    return true;
-                }
-            }
-            return false;
-        }
-    }
-
-    public enum PodPhase {
-        PENDING("Pending"),
-        RUNNING("Running"),
-        SUCCEEDED("Succeeded"),
-        FAILED("Failed"),
-        UNKNOWN("Unknown");
-
-        private final String value;
-
-        PodPhase(String value) {
-            this.value = value;
-        }
-
-        public String value() {
-            return value;
-        }
-    }
-}
