--- src/main/java/io/kestra/plugin/kubernetes/watchers/PodWatcher.java
@@ -1,22 +1,0 @@
-package io.kestra.plugin.kubernetes.watchers;
-
-import org.slf4j.Logger;
-
-import io.fabric8.kubernetes.api.model.Pod;
-
-public class PodWatcher extends AbstractWatch<Pod> {
-    public PodWatcher(Logger logger) {
-        super(logger);
-    }
-
-    protected String logContext(Pod resource) {
-        return String.join(
-            ", ",
-            "Type: " + resource.getClass().getSimpleName(),
-            "Namespace: " + resource.getMetadata().getNamespace(),
-            "Name: " + resource.getMetadata().getName(),
-            "Uid: " + resource.getMetadata().getUid(),
-            "Phase: " + resource.getStatus().getPhase()
-        );
-    }
-}
